# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.CreateTimeBlockRequest:
    properties:
      date:
        type: string
      end_time:
        type: string
      start_time:
        type: string
      task_id:
        type: string
    required:
    - date
    - end_time
    - start_time
    type: object
...
  models.TimeBlock:
    properties:
      completed:
        type: boolean
      created_at:
        type: string
      date:
        type: string
      end_time:
        type: string
      id:
        type: string
      start_time:
        type: string
      task_id:
        type: string
      user_id:
        type: string
    type: object
...
  models.UpdateTimeBlockRequest:
    properties:
      completed:
        type: boolean
      end_time:
        type: string
      start_time:
        type: string
      task_id:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API for task management with Kanban + Timeboxing + AI commands
  title: FLOWSTATE API
  version: "1.0"
paths:
...
  /timeblocks:
    get:
      description: Returns user's time blocks, optionally filtered by date
      parameters:
      - description: Filter by date (YYYY-MM-DD)
        in: query
        name: date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                $ref: '#/definitions/models.TimeBlock'
              type: array
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List time blocks
      tags:
      - timeblocks
    post:
      consumes:
      - application/json
      description: Creates a new time block for the user
      parameters:
      - description: Time block data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.CreateTimeBlockRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.TimeBlock'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create time block
      tags:
      - timeblocks
  /timeblocks/{id}:
    delete:
      description: Deletes a time block by ID
      parameters:
      - description: Time block ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete time block
      tags:
      - timeblocks
    put:
      consumes:
      - application/json
      description: Updates an existing time block
      parameters:
      - description: Time block ID
        in: path
        name: id
        required: true
        type: string
      - description: Data to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.UpdateTimeBlockRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.TimeBlock'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update time block
      tags:
      - timeblocks
  /timeblocks/{id}/complete:
    post:
      description: Marks a time block as completed
      parameters:
      - description: Time block ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.TimeBlock'
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Complete time block
      tags:
      - timeblocks
  /timeblocks/{id}/start:
    post:
      description: Marks a time block as active (in progress)
      parameters:
      - description: Time block ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.TimeBlock'
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Start time block
      tags:
      - timeblocks
  /timeblocks/today:
    get:
      description: Returns today's time blocks for the authenticated user. Includes
        completed blocks with time metrics.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get today's time blocks
      tags:
      - timeblocks
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"